+++ lib/dsc-lib-jsonschema/tests/integration/dsc_repo/mod.rs
@@ -2,3 +2,4 @@
 // Licensed under the MIT License.
 
 #[cfg(test)] mod derive_dsc_repo_schema;
+#[cfg(test)] mod recognized_schema_version;
